requirements-dev.txt
python-decouple==3.8
django-address
phonenumber_field